Added check if the source file exists

+++ src/app.js
@@ -5,12 +5,15 @@ const fs = require('fs');
 
 function copyFile() {
   terminal.question('Enter your command: ', (key) => {
-    const [command, firstFile, secondFile] = key.split(' ');
+    const [command, firstFile, secondFile] = key.trim().split(' ');
 
-    if (command === 'cp' && !fs.existsSync(secondFile)) {
+    if (command === 'cp'
+      && fs.existsSync(firstFile)
+      && !fs.existsSync(secondFile)
+      ) {
       fs.copyFile(firstFile, secondFile, (err) => {
         if (err) {
-          throw new Error('Error!');
+          throw new Error(err);
         }
 
         // eslint-disable-next-line no-console
